+++ src/dotnet/ReSharperPlugin.AlphabetizeIt/SortContextAction.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using JetBrains.ReSharper.Feature.Services.Bulbs;
 using JetBrains.ReSharper.Feature.Services.ContextActions;
 using JetBrains.ReSharper.Feature.Services.CSharp.ContextActions;
@@ -6,15 +5,16 @@ using JetBrains.ReSharper.Feature.Services.Intentions;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.Util;
 using ReSharperPlugin.AlphabetizeIt.Actions;
+using System.Collections.Generic;
 
 namespace ReSharperPlugin.AlphabetizeIt;
 
 [ContextAction(Description = "AlphabetizeIt", GroupType = typeof(CSharpContextActions), Name = "AlphabetizeIt", Priority = 1)]
-public sealed class SortPropertiesContextAction : IContextAction
+public sealed class SortContextAction : IContextAction
 {
     private readonly ICSharpContextActionDataProvider _dataProvider;
 
-    public SortPropertiesContextAction(ICSharpContextActionDataProvider dataProvider)
+    public SortContextAction(ICSharpContextActionDataProvider dataProvider)
     {
         _dataProvider = dataProvider;
     }
@@ -36,6 +36,11 @@ public sealed class SortPropertiesContextAction : IContextAction
         if (!isObjInitializer && isClass && classDec.PropertyDeclarations.Count > 1)
         {
             bulbItems.Add(new SortClassPropertiesAction(_dataProvider));
+        }
+
+        if (!isObjInitializer && isClass && classDec.MethodDeclarations.Count > 1)
+        {
+            bulbItems.Add(new SortClassMethodsAction(_dataProvider));
         }
 
         return bulbItems.ToContextActionIntentions();